fix `enhance_sanity_check` for `PythonPackage`
When trying to add, not replace paths/commands for the sanity check of
Python packages the default `lib/python/%(pyshortver)s/site-packages`
gets replaced.
This leads to duplicating this in many easyconfigs.
Use the `custom_paths` argument for that but still honor if that is
given from a child class.