lammps: select best performing FFT lib
From v `29Aug2024` it is possible to specify the FFT library used in Kokkos by the Cmake setting `FFT_KOKKOS`.
LAMMPS is shipped with the internal KISS FFT library, which is selected by default.
However, for CUDA builds, messages appear at runtime advising the user to use `cuFFT` for better performance.
For CPU-only builds, it is also advised to use FFTW or MKL, if available.