Allow using PythonPackage as extenion in Conda easyblock
This fixes a regression since introduction of `module_load_environment`:
The PythonPackage easyblock tries to set the `type` attribute of 2
module variables that are explicitely removed by the Conda easyblock
leading to a `KeyError`.
So this is expected when used as an extension and we can safely ignore
this error because the type of a variable, that will not be set, doesn't matter.